Changed return value and variables to BigDecimal for better precision

+++ src/main/java/pl/dawidbasa/crediAnalyser/Credit/CreditService.java
@@ -1,5 +1,6 @@
 package pl.dawidbasa.crediAnalyser.Credit;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -9,8 +10,8 @@ public interface CreditService {
 	public void saveCredit(Credit credit);
 	List<Credit> findAll();
 	public List<Credit> sortCreditsByCreditMargin(List<Credit> credits);
-	public List<Double> calculateAllDecreasingInstalments(Credit credit);
-	public Map<String, Double> calculateDecrasingInstalmentDetails(Credit credit);
-	public Map<String, Double> calculateConstantInstalmentDetails(Credit credit);
+	public List<BigDecimal> calculateAllDecreasingInstalments(Credit credit);
+	public Map<String, BigDecimal> calculateDecrasingInstalmentDetails(Credit credit);
+	public Map<String, BigDecimal> calculateConstantInstalmentDetails(Credit credit);
 	
 }
